+++ src/pages/DataProduct/DataProductPage.test.tsx
@@ -2,7 +2,7 @@ import { describe, test } from 'vitest';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { StoreProvider } from '@ska-telescope/ska-gui-local-storage';
-import SdpDataPage from './SdpDataPage';
+import SdpDataPage from './DataProductPage';
 import { AppFlowProvider } from '@/utils/appFlow/AppFlowContext';
 
 const wrapper = (component: React.ReactElement) => {