+++ src/services/axios/get/getCoordinates/getCoordinates.tsx
@@ -64,9 +64,7 @@ const mapping = (
 async function GetCoordinates(targetName: string, skyUnits: number) {
   const units = skyUnits < 0 || skyUnits > MOCK_UNITS.length - 1 ? 0 : skyUnits;
   if (USE_LOCAL_DATA) {
-    return targetName?.trim() === 'M1'
-      ? mapping(MOCK_RESULTS[units])
-      : { error: 'resolve.error.name' };
+    return targetName?.trim() === 'M1' ? mapping(MOCK_RESULTS[units]) : { error: 'name' };
   }
 
   try {