+++ src/utils/validation/validation.tsx
@@ -1,4 +1,14 @@
-import { STATUS_ERROR, STATUS_OK, STATUS_PARTIAL } from './../constants';
+import {
+  MOCK_CALL,
+  MODE_CONTINUUM,
+  MODE_PST,
+  MODE_ZOOM,
+  PAGE_DATA_PRODUCTS,
+  PAGE_OBSERVATION,
+  STATUS_ERROR,
+  STATUS_OK,
+  STATUS_PARTIAL
+} from './../constants';
 import Proposal from './../types/proposal';
 
 export const validateTitlePage = (proposal: Proposal) => {
@@ -15,7 +25,7 @@ export const validateTitlePage = (proposal: Proposal) => {
 
 export const validateTeamPage = (proposal: Proposal) => {
   const result = [STATUS_ERROR, STATUS_OK];
-  const count = proposal.investigators?.length > 0 ? 1 : 0;
+  const count = Array.isArray(proposal.investigators) && proposal.investigators.length > 0 ? 1 : 0;
   return result[count];
 };
 
@@ -23,10 +33,10 @@ export const validateGeneralPage = (proposal: Proposal) => {
   const result = [STATUS_ERROR, STATUS_PARTIAL, STATUS_OK];
   let count = 0;
 
-  if (proposal?.abstract?.length > 0) {
+  if ((proposal?.abstract ?? '').length > 0) {
     count++;
   }
-  if (proposal?.scienceCategory > 0) {
+  if ((proposal?.scienceCategory ?? 0) > 0) {
     count++;
   }
   return result[count];
@@ -61,7 +71,8 @@ export const validateTechnicalPage = (proposal: Proposal) => {
 
 export const validateSDPPage = (proposal: Proposal) => {
   const result = [STATUS_ERROR, STATUS_OK];
-  const count = proposal.dataProductSDP?.length > 0 ? 1 : 0;
+  const count =
+    Array.isArray(proposal?.dataProductSDP) && proposal.dataProductSDP.length > 0 ? 1 : 0;
   return result[count];
 };
 
@@ -79,7 +90,6 @@ export const validateCalibrationPage = (proposal: Proposal) => {
 export const validateLinkingPage = (proposal: Proposal) => {
   const result = [STATUS_ERROR, STATUS_PARTIAL, STATUS_OK];
   const hasTargetObservations = () => (proposal?.targetObservation?.length ?? 0) > 0;
-
   let count = hasTargetObservations() ? 2 : 0;
   return result[count];
 };
@@ -95,10 +105,21 @@ export const validateProposal = (proposal: Proposal) => {
     validateTechnicalPage(proposal),
     validateSDPPage(proposal),
     validateLinkingPage(proposal),
-    validateCalibrationPage(proposal),
-    validateSRCPage()
+    validateCalibrationPage(proposal)
+    // See SRCNet INACTIVE - validateSRCPage()
   ];
   return results;
+};
+
+export const validateProposalNavigation = (proposal: Proposal, page: number) => {
+  if (MOCK_CALL && (page === PAGE_OBSERVATION || page === PAGE_DATA_PRODUCTS)) {
+    return (
+      proposal.scienceCategory === MODE_CONTINUUM ||
+      proposal.scienceCategory === MODE_PST ||
+      proposal.scienceCategory === MODE_ZOOM
+    );
+  }
+  return true;
 };
 
 export function validateSkyDirection1Text(value: string): boolean {